[APPC-1561] Refactored files into more specific packages
