+++ android-appcoins-billing/src/main/java/com/appcoins/sdk/billing/AppcoinsBillingClient.java
@@ -2,6 +2,9 @@ package com.appcoins.sdk.billing;
 
 import android.app.Activity;
 import android.content.Intent;
+import com.appcoins.sdk.billing.listeners.AppCoinsBillingStateListener;
+import com.appcoins.sdk.billing.listeners.ConsumeResponseListener;
+import com.appcoins.sdk.billing.listeners.SkuDetailsResponseListener;
 
 public interface AppcoinsBillingClient {
   PurchasesResult queryPurchases(String skuType);
