+++ android-appcoins-billing/src/main/java/com/appcoins/sdk/billing/ConnectionLifeCycle.java
@@ -1,6 +1,7 @@
 package com.appcoins.sdk.billing;
 
 import android.os.IBinder;
+import com.appcoins.sdk.billing.listeners.AppCoinsBillingStateListener;
 
 public interface ConnectionLifeCycle {
   void onConnect(IBinder service, final AppCoinsBillingStateListener listener);
