+++ android-appcoins-billing/src/main/java/com/appcoins/sdk/billing/RepositoryConnection.java
@@ -1,5 +1,7 @@
 package com.appcoins.sdk.billing;
 
+import com.appcoins.sdk.billing.listeners.AppCoinsBillingStateListener;
+
 public interface RepositoryConnection {
   void startConnection(final AppCoinsBillingStateListener listener);
