+++ android-appcoins-billing/src/main/java/com/appcoins/sdk/billing/service/WalletBillingService.java
@@ -4,7 +4,6 @@ import android.os.Bundle;
 import android.os.IBinder;
 import android.os.RemoteException;
 import com.appcoins.billing.AppcoinsBilling;
-import com.appcoins.sdk.billing.AppCoinsBillingStateListener;
 import com.appcoins.sdk.billing.helpers.AppcoinsBillingStubHelper;
 
 public class WalletBillingService implements AppcoinsBilling {
