+++ appcoins-billing/src/main/java/com/appcoins/sdk/billing/Billing.java
@@ -1,5 +1,9 @@
 package com.appcoins.sdk.billing;
 
+import com.appcoins.sdk.billing.exceptions.ServiceConnectionException;
+import com.appcoins.sdk.billing.listeners.ConsumeResponseListener;
+import com.appcoins.sdk.billing.listeners.SkuDetailsResponseListener;
+
 public interface Billing {
 
   PurchasesResult queryPurchases(String skuType);
