+++ appcoins-billing/src/main/java/com/appcoins/sdk/billing/exceptions/ServiceConnectionException.java
@@ -1,4 +1,4 @@
-package com.appcoins.sdk.billing;
+package com.appcoins.sdk.billing.exceptions;
 
 public class ServiceConnectionException extends Exception {
