+++ appcoins-billing/src/main/java/com/appcoins/sdk/billing/listeners/SkuDetailsResponseListener.java
@@ -1,5 +1,6 @@
-package com.appcoins.sdk.billing;
+package com.appcoins.sdk.billing.listeners;
 
+import com.appcoins.sdk.billing.SkuDetails;
 import java.util.List;
 
 public interface SkuDetailsResponseListener {
